requirements.txt
Flask==3.1.0
python-dotenv==1.1.0
langchain==0.3.24
langchain-core==0.3.55
langchain-community==0.3.22
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.8
sentence-transformers==4.1.0
transformers==4.51.3
faiss-cpu==1.10.0
google-generativeai==0.8.5
unstructured==0.17.2
unstructured-client==0.33.1
huggingface-hub==0.30.2
torch==2.6.0
scikit-learn==1.6.1
nltk==3.9.1




